Remove node polyfills from renderer
