+++ package.json
@@ -115,12 +115,9 @@
     "@react-nano/use-event-source": "^0.12.0",
     "@vitejs/plugin-react-swc": "^3.6.0",
     "adm-zip": "^0.5.10",
-    "assert": "^2.1.0",
     "bezier-js": "^6.1.4",
-    "buffer": "^6.0.3",
     "chakra-ui-markdown-renderer": "^4.1.0",
     "chokidar": "^3.5.3",
-    "crypto-browserify": "^3.12.0",
     "decompress": "^4.2.1",
     "electron-log": "^5.1.2",
     "electron-squirrel-startup": "^1.0.0",
@@ -129,7 +126,6 @@
     "fast-deep-equal": "^3.1.3",
     "framer-motion": "^6.3.6",
     "html-to-image": "^1.10.8",
-    "https-browserify": "^1.0.0",
     "i18next": "^22.0.5",
     "i18next-http-backend": "^2.5.0",
     "javascript-time-ago": "^2.5.9",
@@ -137,10 +133,8 @@
     "minisearch": "^6.1.0",
     "node-localstorage": "^2.2.1",
     "nodejs-file-downloader": "^4.9.3",
-    "os-browserify": "^0.3.0",
     "os-utils": "^0.0.14",
     "portfinder": "^1.0.32",
-    "process": "^0.11.10",
     "re-resizable": "^6.9.9",
     "react": "^18.1.0",
     "react-colorful": "^5.6.1",
@@ -155,13 +149,9 @@
     "reactflow": "^11.10.1",
     "rregex": "^1.10.11",
     "semver": "^7.3.7",
-    "stream-browserify": "^3.0.0",
-    "stream-http": "^3.2.0",
     "systeminformation": "^5.21.7",
-    "url": "^0.11.3",
     "use-context-selector": "^1.4.0",
     "use-debounce": "^8.0.1",
-    "util": "^0.12.5",
     "uuid": "^8.3.2",
     "vite": "^5.2.6",
     "wcag-contrast": "^3.0.0",
